// src/app/shared/models/shippingParameterDetails.ts
export class ShippingParameterDetails {
    DSPNumber: any;
    SiteId: Number;
    MaterialId: string;
    ManufacturingDateTimeFrom: any;
    ManufacturingDateTimeThru: any;
    PalletIdFrom: string;
    PalletIdThru: string;
    Comments: string;
    Crews: Array<string>;
    Shifts: Array<string>;
    Lines: Array<string>;
    Sites: Array<string>;
    Type: string;
    Status: string;
}